Refactor constants to package
