+++ src/main/java/constants/ErrorConstant.java
@@ -1,4 +1,4 @@
-package utility;
+package constants;
 
 public class ErrorConstant {
