+++ src/main/java/constants/HealthConstant.java
@@ -1,4 +1,4 @@
-package utility;
+package constants;
 
 public class HealthConstant {
