+++ src/main/java/ui/Handler.java
@@ -3,10 +3,10 @@ package ui;
 import health.HealthList;
 import storage.DataFile;
 import utility.CustomExceptions;
-import utility.ErrorConstant;
-import utility.UiConstant;
-import utility.HealthConstant;
-import utility.WorkoutConstant;
+import constants.ErrorConstant;
+import constants.UiConstant;
+import constants.HealthConstant;
+import constants.WorkoutConstant;
 import utility.Command;
 import utility.Filters;
 import utility.Parser;
