+++ src/main/java/workouts/Run.java
@@ -4,9 +4,9 @@ import java.time.LocalDate;
 
 import utility.CustomExceptions;
 import utility.Parser;
-import utility.ErrorConstant;
-import utility.UiConstant;
-import utility.WorkoutConstant;
+import constants.ErrorConstant;
+import constants.UiConstant;
+import constants.WorkoutConstant;
 
 
 /**
