+++ src/test/java/health/PeriodTest.java
@@ -5,7 +5,7 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
 import utility.CustomExceptions;
-import utility.HealthConstant;
+import constants.HealthConstant;
 
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
